Added Premake script to create consistent project files on all supported platfroms & some cleanup

+++ source/main.cpp
@@ -86,7 +86,7 @@ int main(int argc, char ** argv)
 
 	//HashTestBuffer();
 	
-	int a = strlen(argv[1]);
+	u64 a = strlen(argv[1]);
 	String filename = {(u8*)argv[1],  strlen(argv[1])};
 	parse_file(filename);
